LRQA-20643 Merge MobileDriverToSeleniumBridge into BaseMobileDriverImpl
