+++ modules/test/poshi-runner/src/main/java/com/liferay/poshi/runner/selenium/BaseMobileDriverImpl.java
@@ -16,19 +16,67 @@ package com.liferay.poshi.runner.selenium;
 
 import com.liferay.poshi.runner.util.PropsValues;
 
+import java.util.List;
+
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.WebElement;
+
+import org.w3c.dom.Node;
 
 /**
  * @author Kenji Heigel
  */
 public abstract class BaseMobileDriverImpl
-	extends MobileDriverToSeleniumBridge implements LiferaySelenium {
+	extends MobileDriverWrapper implements LiferaySelenium {
 
 	public BaseMobileDriverImpl(String browserURL, WebDriver webDriver) {
 		super(webDriver);
 
+		WebDriverHelper.setDefaultWindowHandle(webDriver.getWindowHandle());
+
 		System.setProperty("java.awt.headless", "false");
+	}
+
+	@Override
+	public void addCustomRequestHeader(String key, String value) {
+		throw new UnsupportedOperationException();
+	}
+
+	@Override
+	public void addLocationStrategy(
+		String strategyName, String functionDefinition) {
+
+		throw new UnsupportedOperationException();
+	}
+
+	@Override
+	public void addScript(String script, String scriptTagId) {
+		throw new UnsupportedOperationException();
+	}
+
+	@Override
+	public void addSelection(String locator, String optionLocator) {
+		WebDriverHelper.addSelection(this, locator, optionLocator);
+	}
+
+	@Override
+	public void allowNativeXpath(String allow) {
+		throw new UnsupportedOperationException();
+	}
+
+	@Override
+	public void altKeyDown() {
+		throw new UnsupportedOperationException();
+	}
+
+	@Override
+	public void altKeyUp() {
+		throw new UnsupportedOperationException();
+	}
+
+	@Override
+	public void answerOnNextPrompt(String answer) {
+		throw new UnsupportedOperationException();
 	}
 
 	@Override
@@ -240,17 +288,97 @@ public abstract class BaseMobileDriverImpl
 	}
 
 	@Override
-	public void clickAndWait(String locator) {
-		super.click(locator);
+	public void assignId(String locator, String identifier) {
+		throw new UnsupportedOperationException();
+	}
 
-		super.waitForPageToLoad("30000");
+	@Override
+	public void attachFile(String fieldLocator, String fileLocator) {
+		throw new UnsupportedOperationException();
+	}
+
+	@Override
+	public void captureEntirePageScreenshot(String fileName, String kwargs) {
+		throw new UnsupportedOperationException();
+	}
+
+	@Override
+	public String captureEntirePageScreenshotToString(String kwargs) {
+		throw new UnsupportedOperationException();
+	}
+
+	@Override
+	public String captureNetworkTraffic(String type) {
+		throw new UnsupportedOperationException();
+	}
+
+	@Override
+	public void captureScreenshot(String fileName) {
+		throw new UnsupportedOperationException();
+	}
+
+	@Override
+	public String captureScreenshotToString() {
+		throw new UnsupportedOperationException();
+	}
+
+	@Override
+	public void check(String locator) {
+		WebDriverHelper.check(this, locator);
+	}
+
+	@Override
+	public void chooseCancelOnNextConfirmation() {
+		throw new UnsupportedOperationException();
+	}
+
+	@Override
+	public void chooseOkOnNextConfirmation() {
+		throw new UnsupportedOperationException();
+	}
+
+	@Override
+	public void click(String locator) {
+		try {
+			tap(locator);
+		}
+		catch (Exception e) {
+			if (!isInViewport(locator)) {
+				swipeWebElementIntoView(locator);
+			}
+
+			tap(locator);
+		}
+	}
+
+	@Override
+	public void clickAndWait(String locator) {
+		click(locator);
+
+		waitForPageToLoad("30000");
+	}
+
+	@Override
+	public void clickAt(String locator, String coordString) {
+		clickAt(locator, coordString, true);
+	}
+
+	public void clickAt(
+		String locator, String coordString, boolean scrollIntoView) {
+
+		click(locator);
 	}
 
 	@Override
 	public void clickAtAndWait(String locator, String coordString) {
-		super.clickAt(locator, coordString);
+		clickAt(locator, coordString);
 
-		super.waitForPageToLoad("30000");
+		waitForPageToLoad("30000");
+	}
+
+	@Override
+	public void close() {
+		super.close();
 	}
 
 	@Override
@@ -259,6 +387,26 @@ public abstract class BaseMobileDriverImpl
 
 		LiferaySeleniumHelper.connectToEmailAccount(
 			emailAddress, emailPassword);
+	}
+
+	@Override
+	public void contextMenu(String locator) {
+		throw new UnsupportedOperationException();
+	}
+
+	@Override
+	public void contextMenuAt(String locator, String coordString) {
+		throw new UnsupportedOperationException();
+	}
+
+	@Override
+	public void controlKeyDown() {
+		throw new UnsupportedOperationException();
+	}
+
+	@Override
+	public void controlKeyUp() {
+		throw new UnsupportedOperationException();
 	}
 
 	@Override
@@ -272,8 +420,56 @@ public abstract class BaseMobileDriverImpl
 	}
 
 	@Override
+	public void createCookie(String nameValuePair, String optionsString) {
+		throw new UnsupportedOperationException();
+	}
+
+	@Override
 	public void deleteAllEmails() throws Exception {
 		LiferaySeleniumHelper.deleteAllEmails();
+	}
+
+	@Override
+	public void deleteAllVisibleCookies() {
+		throw new UnsupportedOperationException();
+	}
+
+	@Override
+	public void deleteCookie(String name, String optionsString) {
+		throw new UnsupportedOperationException();
+	}
+
+	@Override
+	public void deselectPopUp() {
+		throw new UnsupportedOperationException();
+	}
+
+	@Override
+	public void doubleClick(String locator) {
+		throw new UnsupportedOperationException();
+	}
+
+	@Override
+	public void doubleClickAt(String locator, String coordString) {
+		throw new UnsupportedOperationException();
+	}
+
+	@Override
+	public void dragAndDrop(String locator, String coordString) {
+		throw new UnsupportedOperationException();
+	}
+
+	@Override
+	public void dragAndDropToObject(
+		String locatorOfObjectToBeDragged,
+		String locatorOfDragDestinationObject) {
+
+		throw new UnsupportedOperationException();
+	}
+
+	@Override
+	public void dragdrop(String locator, String movementsString) {
+		throw new UnsupportedOperationException();
 	}
 
 	@Override
@@ -284,6 +480,86 @@ public abstract class BaseMobileDriverImpl
 	@Override
 	public void fail(String message) {
 		LiferaySeleniumHelper.fail(message);
+	}
+
+	@Override
+	public void fireEvent(String locator, String eventName) {
+		throw new UnsupportedOperationException();
+	}
+
+	@Override
+	public void focus(String locator) {
+		throw new UnsupportedOperationException();
+	}
+
+	@Override
+	public String getAlert() {
+		throw new UnsupportedOperationException();
+	}
+
+	@Override
+	public String[] getAllButtons() {
+		throw new UnsupportedOperationException();
+	}
+
+	@Override
+	public String[] getAllFields() {
+		throw new UnsupportedOperationException();
+	}
+
+	@Override
+	public String[] getAllLinks() {
+		throw new UnsupportedOperationException();
+	}
+
+	@Override
+	public String[] getAllWindowIds() {
+		throw new UnsupportedOperationException();
+	}
+
+	@Override
+	public String[] getAllWindowNames() {
+		throw new UnsupportedOperationException();
+	}
+
+	@Override
+	public String[] getAllWindowTitles() {
+		throw new UnsupportedOperationException();
+	}
+
+	@Override
+	public String getAttribute(String attributeLocator) {
+		return WebDriverHelper.getAttribute(this, attributeLocator);
+	}
+
+	@Override
+	public String[] getAttributeFromAllWindows(String attributeName) {
+		throw new UnsupportedOperationException();
+	}
+
+	@Override
+	public String getBodyText() {
+		throw new UnsupportedOperationException();
+	}
+
+	@Override
+	public String getConfirmation() {
+		return WebDriverHelper.getConfirmation(this);
+	}
+
+	@Override
+	public String getCookie() {
+		throw new UnsupportedOperationException();
+	}
+
+	@Override
+	public String getCookieByName(String name) {
+		throw new UnsupportedOperationException();
+	}
+
+	@Override
+	public Number getCssCount(String css) {
+		throw new UnsupportedOperationException();
 	}
 
 	@Override
@@ -303,6 +579,31 @@ public abstract class BaseMobileDriverImpl
 
 	@Override
 	public String getCurrentYear() {
+		throw new UnsupportedOperationException();
+	}
+
+	@Override
+	public Number getCursorPosition(String locator) {
+		throw new UnsupportedOperationException();
+	}
+
+	@Override
+	public Number getElementHeight(String locator) {
+		throw new UnsupportedOperationException();
+	}
+
+	@Override
+	public Number getElementIndex(String locator) {
+		throw new UnsupportedOperationException();
+	}
+
+	@Override
+	public Number getElementPositionLeft(String locator) {
+		throw new UnsupportedOperationException();
+	}
+
+	@Override
+	public Number getElementPositionTop(String locator) {
 		throw new UnsupportedOperationException();
 	}
 
@@ -355,6 +656,11 @@ public abstract class BaseMobileDriverImpl
 	}
 
 	@Override
+	public Number getElementWidth(String locator) {
+		throw new UnsupportedOperationException();
+	}
+
+	@Override
 	public String getEmailBody(String index) throws Exception {
 		return LiferaySeleniumHelper.getEmailBody(index);
 	}
@@ -365,12 +671,54 @@ public abstract class BaseMobileDriverImpl
 	}
 
 	@Override
+	public String getEval(String script) {
+		return WebDriverHelper.getEval(this, script);
+	}
+
+	@Override
+	public String getExpression(String expression) {
+		throw new UnsupportedOperationException();
+	}
+
+	@Override
 	public String getFirstNumber(String locator) {
 		throw new UnsupportedOperationException();
 	}
 
 	@Override
 	public String getFirstNumberIncrement(String locator) {
+		throw new UnsupportedOperationException();
+	}
+
+	public Node getHtmlNode(String locator) {
+		throw new UnsupportedOperationException();
+	}
+
+	public String getHtmlNodeHref(String locator) {
+		throw new UnsupportedOperationException();
+	}
+
+	public String getHtmlNodeText(String locator) {
+		throw new UnsupportedOperationException();
+	}
+
+	@Override
+	public String getHtmlSource() {
+		throw new UnsupportedOperationException();
+	}
+
+	@Override
+	public String getLocation() {
+		return WebDriverHelper.getLocation(this);
+	}
+
+	@Override
+	public String getLog() {
+		throw new UnsupportedOperationException();
+	}
+
+	@Override
+	public Number getMouseSpeed() {
 		throw new UnsupportedOperationException();
 	}
 
@@ -395,8 +743,72 @@ public abstract class BaseMobileDriverImpl
 	}
 
 	@Override
+	public String getPrompt() {
+		throw new UnsupportedOperationException();
+	}
+
+	@Override
+	public String getSelectedId(String selectLocator) {
+		throw new UnsupportedOperationException();
+	}
+
+	@Override
+	public String[] getSelectedIds(String selectLocator) {
+		throw new UnsupportedOperationException();
+	}
+
+	@Override
+	public String getSelectedIndex(String selectLocator) {
+		throw new UnsupportedOperationException();
+	}
+
+	@Override
+	public String[] getSelectedIndexes(String selectLocator) {
+		throw new UnsupportedOperationException();
+	}
+
+	@Override
+	public String getSelectedLabel(String selectLocator) {
+		return getSelectedLabel(selectLocator, null);
+	}
+
+	public String getSelectedLabel(String selectLocator, String timeout) {
+		return WebDriverHelper.getSelectedLabel(this, selectLocator, timeout);
+	}
+
+	@Override
+	public String[] getSelectedLabels(String selectLocator) {
+		return WebDriverHelper.getSelectedLabels(this, selectLocator);
+	}
+
+	@Override
+	public String getSelectedValue(String selectLocator) {
+		throw new UnsupportedOperationException();
+	}
+
+	@Override
+	public String[] getSelectedValues(String selectLocator) {
+		throw new UnsupportedOperationException();
+	}
+
+	@Override
+	public String[] getSelectOptions(String selectLocator) {
+		throw new UnsupportedOperationException();
+	}
+
+	@Override
 	public String getSikuliImagesDirName() {
 		return _SIKULI_IMAGES_DIR_NAME;
+	}
+
+	@Override
+	public String getSpeed() {
+		throw new UnsupportedOperationException();
+	}
+
+	@Override
+	public String getTable(String tableCellAddress) {
+		throw new UnsupportedOperationException();
 	}
 
 	@Override
@@ -405,10 +817,79 @@ public abstract class BaseMobileDriverImpl
 	}
 
 	@Override
-	public void goBackAndWait() {
-		super.goBack();
+	public String getText(String locator) {
+		throw new UnsupportedOperationException();
+	}
 
-		super.waitForPageToLoad("30000");
+	@Override
+	public String getTitle() {
+		return super.getTitle();
+	}
+
+	@Override
+	public String getValue(String locator) {
+		throw new UnsupportedOperationException();
+	}
+
+	@Override
+	public boolean getWhetherThisFrameMatchFrameExpression(
+		String currentFrameString, String target) {
+
+		throw new UnsupportedOperationException();
+	}
+
+	@Override
+	public boolean getWhetherThisWindowMatchWindowExpression(
+		String currentWindowString, String target) {
+
+		throw new UnsupportedOperationException();
+	}
+
+	@Override
+	public Number getXpathCount(String xPath) {
+		throw new UnsupportedOperationException();
+	}
+
+	@Override
+	public void goBack() {
+		WebDriverHelper.goBack(this);
+	}
+
+	@Override
+	public void goBackAndWait() {
+		goBack();
+
+		waitForPageToLoad("30000");
+	}
+
+	@Override
+	public void highlight(String locator) {
+		throw new UnsupportedOperationException();
+	}
+
+	@Override
+	public void ignoreAttributesWithoutValue(String ignore) {
+		throw new UnsupportedOperationException();
+	}
+
+	@Override
+	public boolean isAlertPresent() {
+		throw new UnsupportedOperationException();
+	}
+
+	@Override
+	public boolean isChecked(String locator) {
+		WebElement webElement = getWebElement(locator, "1");
+
+		if (!webElement.isDisplayed()) {
+			return webElement.isDisplayed();
+		}
+
+		if (!isInViewport(locator)) {
+			swipeWebElementIntoView(locator);
+		}
+
+		return webElement.isSelected();
 	}
 
 	@Override
@@ -417,8 +898,28 @@ public abstract class BaseMobileDriverImpl
 	}
 
 	@Override
+	public boolean isConfirmationPresent() {
+		throw new UnsupportedOperationException();
+	}
+
+	@Override
+	public boolean isCookiePresent(String name) {
+		throw new UnsupportedOperationException();
+	}
+
+	@Override
+	public boolean isEditable(String locator) {
+		throw new UnsupportedOperationException();
+	}
+
+	@Override
 	public boolean isElementNotPresent(String locator) {
 		return WebDriverHelper.isElementNotPresent(this, locator);
+	}
+
+	@Override
+	public boolean isElementPresent(String locator) {
+		return WebDriverHelper.isElementPresent(this, locator);
 	}
 
 	@Override
@@ -429,6 +930,25 @@ public abstract class BaseMobileDriverImpl
 	@Override
 	public boolean isHTMLSourceTextPresent(String value) throws Exception {
 		return LiferaySeleniumHelper.isHTMLSourceTextPresent(this, value);
+	}
+
+	public boolean isInViewport(String locator) {
+		int elementPositionCenterY = WebDriverHelper.getElementPositionCenterY(
+			this, locator);
+
+		int viewportPositionBottom = WebDriverHelper.getViewportPositionBottom(
+			this);
+
+		int viewportPositionTop = WebDriverHelper.getScrollOffsetY(this);
+
+		if ((elementPositionCenterY >= viewportPositionBottom) ||
+			(elementPositionCenterY <= viewportPositionTop)) {
+
+			return false;
+		}
+		else {
+			return true;
+		}
 	}
 
 	@Override
@@ -467,8 +987,18 @@ public abstract class BaseMobileDriverImpl
 	}
 
 	@Override
+	public boolean isOrdered(String locator1, String locator2) {
+		throw new UnsupportedOperationException();
+	}
+
+	@Override
 	public boolean isPartialText(String locator, String value) {
 		return WebDriverHelper.isPartialText(this, locator, value);
+	}
+
+	@Override
+	public boolean isPromptPresent() {
+		throw new UnsupportedOperationException();
 	}
 
 	@Override
@@ -479,6 +1009,11 @@ public abstract class BaseMobileDriverImpl
 	@Override
 	public boolean isSikuliImagePresent(String image) throws Exception {
 		return LiferaySeleniumHelper.isSikuliImagePresent(this, image);
+	}
+
+	@Override
+	public boolean isSomethingSelected(String selectLocator) {
+		throw new UnsupportedOperationException();
 	}
 
 	@Override
@@ -497,8 +1032,33 @@ public abstract class BaseMobileDriverImpl
 	}
 
 	@Override
+	public boolean isTextPresent(String pattern) {
+		throw new UnsupportedOperationException();
+	}
+
+	@Override
 	public boolean isValue(String locator, String value) throws Exception {
 		return value.equals(getElementValue(locator, "1"));
+	}
+
+	@Override
+	public boolean isVisible(String locator) {
+		WebElement webElement = getWebElement(locator, "1");
+
+		if (PropsValues.BROWSER_TYPE.equals("android")) {
+			if (!isInViewport(locator)) {
+				swipeWebElementIntoView(locator);
+			}
+
+			return isInViewport(locator);
+		}
+		else {
+			if (!webElement.isDisplayed()) {
+				WebDriverHelper.scrollWebElementIntoView(this, webElement);
+			}
+
+			return webElement.isDisplayed();
+		}
 	}
 
 	@Override
@@ -512,7 +1072,22 @@ public abstract class BaseMobileDriverImpl
 	}
 
 	@Override
+	public void keyDown(String locator, String keySequence) {
+		throw new UnsupportedOperationException();
+	}
+
+	@Override
 	public void keyDownAndWait(String locator, String keySequence) {
+		throw new UnsupportedOperationException();
+	}
+
+	@Override
+	public void keyDownNative(String keycode) {
+		throw new UnsupportedOperationException();
+	}
+
+	@Override
+	public void keyPress(String locator, String keySequence) {
 		throw new UnsupportedOperationException();
 	}
 
@@ -522,7 +1097,22 @@ public abstract class BaseMobileDriverImpl
 	}
 
 	@Override
+	public void keyPressNative(String keycode) {
+		throw new UnsupportedOperationException();
+	}
+
+	@Override
+	public void keyUp(String locator, String keySequence) {
+		throw new UnsupportedOperationException();
+	}
+
+	@Override
 	public void keyUpAndWait(String locator, String keySequence) {
+		throw new UnsupportedOperationException();
+	}
+
+	@Override
+	public void keyUpNative(String keycode) {
 		throw new UnsupportedOperationException();
 	}
 
@@ -532,8 +1122,92 @@ public abstract class BaseMobileDriverImpl
 	}
 
 	@Override
+	public void metaKeyDown() {
+		throw new UnsupportedOperationException();
+	}
+
+	@Override
+	public void metaKeyUp() {
+		throw new UnsupportedOperationException();
+	}
+
+	@Override
+	public void mouseDown(String locator) {
+		throw new UnsupportedOperationException();
+	}
+
+	@Override
+	public void mouseDownAt(String locator, String coordString) {
+		throw new UnsupportedOperationException();
+	}
+
+	@Override
+	public void mouseDownRight(String locator) {
+		throw new UnsupportedOperationException();
+	}
+
+	@Override
+	public void mouseDownRightAt(String locator, String coordString) {
+		throw new UnsupportedOperationException();
+	}
+
+	@Override
+	public void mouseMove(String locator) {
+		throw new UnsupportedOperationException();
+	}
+
+	@Override
+	public void mouseMoveAt(String locator, String coordString) {
+		throw new UnsupportedOperationException();
+	}
+
+	@Override
+	public void mouseOut(String locator) {
+		throw new UnsupportedOperationException();
+	}
+
+	@Override
+	public void mouseOver(String locator) {
+	}
+
+	@Override
 	public void mouseRelease() {
 		throw new UnsupportedOperationException();
+	}
+
+	@Override
+	public void mouseUp(String locator) {
+		throw new UnsupportedOperationException();
+	}
+
+	@Override
+	public void mouseUpAt(String locator, String coordString) {
+		throw new UnsupportedOperationException();
+	}
+
+	@Override
+	public void mouseUpRight(String locator) {
+		throw new UnsupportedOperationException();
+	}
+
+	@Override
+	public void mouseUpRightAt(String locator, String coordString) {
+		throw new UnsupportedOperationException();
+	}
+
+	@Override
+	public void open(String url) {
+		WebDriverHelper.open(this, url);
+	}
+
+	@Override
+	public void open(String url, String ignoreResponseCode) {
+		throw new UnsupportedOperationException();
+	}
+
+	@Override
+	public void openWindow(String url, String windowID) {
+		open(url);
 	}
 
 	@Override
@@ -551,13 +1225,48 @@ public abstract class BaseMobileDriverImpl
 	}
 
 	@Override
+	public void refresh() {
+		WebDriverHelper.refresh(this);
+	}
+
+	@Override
 	public void refreshAndWait() {
+		throw new UnsupportedOperationException();
+	}
+
+	@Override
+	public void removeAllSelections(String locator) {
+		throw new UnsupportedOperationException();
+	}
+
+	@Override
+	public void removeScript(String scriptTagId) {
+		throw new UnsupportedOperationException();
+	}
+
+	@Override
+	public void removeSelection(String locator, String optionLocator) {
 		throw new UnsupportedOperationException();
 	}
 
 	@Override
 	public void replyToEmail(String to, String body) throws Exception {
 		LiferaySeleniumHelper.replyToEmail(this, to, body);
+	}
+
+	@Override
+	public String retrieveLastRemoteControlLogs() {
+		throw new UnsupportedOperationException();
+	}
+
+	@Override
+	public void rollup(String rollupName, String kwargs) {
+		throw new UnsupportedOperationException();
+	}
+
+	@Override
+	public void runScript(String script) {
+		getEval(script);
 	}
 
 	@Override
@@ -590,10 +1299,30 @@ public abstract class BaseMobileDriverImpl
 	}
 
 	@Override
-	public void selectAndWait(String selectLocator, String optionLocator) {
-		super.select(selectLocator, optionLocator);
+	public void select(String selectLocator, String optionLocator) {
+		WebDriverHelper.select(this, selectLocator, optionLocator);
+	}
 
-		super.waitForPageToLoad("30000");
+	@Override
+	public void selectAndWait(String selectLocator, String optionLocator) {
+		select(selectLocator, optionLocator);
+
+		waitForPageToLoad("30000");
+	}
+
+	@Override
+	public void selectFrame(String locator) {
+		WebDriverHelper.selectFrame(this, locator);
+	}
+
+	@Override
+	public void selectPopUp(String windowID) {
+		throw new UnsupportedOperationException();
+	}
+
+	@Override
+	public void selectWindow(String windowID) {
+		WebDriverHelper.selectWindow(this, windowID);
 	}
 
 	@Override
@@ -639,12 +1368,51 @@ public abstract class BaseMobileDriverImpl
 	}
 
 	@Override
+	public void setBrowserLogLevel(String logLevel) {
+		throw new UnsupportedOperationException();
+	}
+
+	@Override
+	public void setContext(String context) {
+		throw new UnsupportedOperationException();
+	}
+
+	@Override
+	public void setCursorPosition(String locator, String position) {
+		throw new UnsupportedOperationException();
+	}
+
+	@Override
 	public void setDefaultTimeout() {
+	}
+
+	public void setDefaultTimeoutImplicit() {
+		WebDriverHelper.setDefaultTimeoutImplicit(this);
+	}
+
+	@Override
+	public void setExtensionJs(String extensionJs) {
+		throw new UnsupportedOperationException();
+	}
+
+	@Override
+	public void setMouseSpeed(String pixels) {
+		throw new UnsupportedOperationException();
 	}
 
 	@Override
 	public void setPrimaryTestSuiteName(String primaryTestSuiteName) {
 		_primaryTestSuiteName = primaryTestSuiteName;
+	}
+
+	@Override
+	public void setSpeed(String value) {
+		throw new UnsupportedOperationException();
+	}
+
+	@Override
+	public void setTimeout(String timeout) {
+		throw new UnsupportedOperationException();
 	}
 
 	@Override
@@ -654,6 +1422,31 @@ public abstract class BaseMobileDriverImpl
 
 	@Override
 	public void setWindowSize(String coordString) {
+		throw new UnsupportedOperationException();
+	}
+
+	@Override
+	public void shiftKeyDown() {
+		throw new UnsupportedOperationException();
+	}
+
+	@Override
+	public void shiftKeyUp() {
+		throw new UnsupportedOperationException();
+	}
+
+	@Override
+	public void showContextualBanner() {
+		throw new UnsupportedOperationException();
+	}
+
+	@Override
+	public void showContextualBanner(String className, String methodName) {
+		throw new UnsupportedOperationException();
+	}
+
+	@Override
+	public void shutDownSeleniumServer() {
 		throw new UnsupportedOperationException();
 	}
 
@@ -738,11 +1531,41 @@ public abstract class BaseMobileDriverImpl
 	}
 
 	@Override
+	public void start() {
+		throw new UnsupportedOperationException();
+	}
+
+	@Override
+	public void start(Object optionsObject) {
+		throw new UnsupportedOperationException();
+	}
+
+	@Override
+	public void start(String optionsString) {
+		throw new UnsupportedOperationException();
+	}
+
+	@Override
 	public void startLogger() {
 	}
 
 	@Override
+	public void stop() {
+		quit();
+	}
+
+	@Override
 	public void stopLogger() {
+	}
+
+	@Override
+	public void submit(String formLocator) {
+		throw new UnsupportedOperationException();
+	}
+
+	@Override
+	public void type(String locator, String value) {
+		WebDriverHelper.type(this, locator, value);
 	}
 
 	@Override
@@ -761,6 +1584,10 @@ public abstract class BaseMobileDriverImpl
 	}
 
 	@Override
+	public void typeKeys(String locator, String value) {
+		throw new UnsupportedOperationException();
+	}
+
 	public void typeKeys(String locator, String value, boolean typeAceEditor) {
 		throw new UnsupportedOperationException();
 	}
@@ -768,6 +1595,11 @@ public abstract class BaseMobileDriverImpl
 	@Override
 	public void typeScreen(String value) {
 		LiferaySeleniumHelper.typeScreen(value);
+	}
+
+	@Override
+	public void uncheck(String locator) {
+		WebDriverHelper.uncheck(this, locator);
 	}
 
 	@Override
@@ -788,6 +1620,16 @@ public abstract class BaseMobileDriverImpl
 	}
 
 	@Override
+	public void useXpathLibrary(String libraryName) {
+		throw new UnsupportedOperationException();
+	}
+
+	@Override
+	public void waitForCondition(String script, String timeout) {
+		throw new UnsupportedOperationException();
+	}
+
+	@Override
 	public void waitForConfirmation(String pattern) throws Exception {
 		LiferaySeleniumHelper.waitForConfirmation(this, pattern);
 	}
@@ -800,6 +1642,11 @@ public abstract class BaseMobileDriverImpl
 	@Override
 	public void waitForElementPresent(String locator) throws Exception {
 		LiferaySeleniumHelper.waitForElementPresent(this, locator);
+	}
+
+	@Override
+	public void waitForFrameToLoad(String frameAddress, String timeout) {
+		throw new UnsupportedOperationException();
 	}
 
 	@Override
@@ -833,10 +1680,20 @@ public abstract class BaseMobileDriverImpl
 	}
 
 	@Override
+	public void waitForPageToLoad(String timeout) {
+		throw new UnsupportedOperationException();
+	}
+
+	@Override
 	public void waitForPartialText(String locator, String value)
 		throws Exception {
 
 		LiferaySeleniumHelper.waitForPartialText(this, locator, value);
+	}
+
+	@Override
+	public void waitForPopUp(String windowID, String timeout) {
+		throw new UnsupportedOperationException();
 	}
 
 	@Override
@@ -873,8 +1730,43 @@ public abstract class BaseMobileDriverImpl
 	}
 
 	@Override
+	public void windowFocus() {
+		throw new UnsupportedOperationException();
+	}
+
+	@Override
+	public void windowMaximize() {
+		throw new UnsupportedOperationException();
+	}
+
+	@Override
 	public void windowMaximizeAndWait() {
 		throw new UnsupportedOperationException();
+	}
+
+	protected WebElement getWebElement(String locator) {
+		return WebDriverHelper.getWebElement(this, locator);
+	}
+
+	protected WebElement getWebElement(String locator, String timeout) {
+		return WebDriverHelper.getWebElement(this, locator, timeout);
+	}
+
+	protected List<WebElement> getWebElements(String locator) {
+		return WebDriverHelper.getWebElements(this, locator);
+	}
+
+	protected List<WebElement> getWebElements(String locator, String timeout) {
+		return WebDriverHelper.getWebElements(this, locator, timeout);
+	}
+
+	protected void swipeWebElementIntoView(String locator) {
+		WebElement webElement = getWebElement(locator, "1");
+
+		WebDriverHelper.scrollWebElementIntoView(this, webElement);
+	}
+
+	protected void tap(String locator) {
 	}
 
 	private static final String _OUTPUT_DIR_NAME = PropsValues.OUTPUT_DIR_NAME;
